refactor: post utils로 분리
