+++ app/feed/posts.tsx
@@ -1,17 +1,6 @@
 import { useEffect, useState, useRef } from 'react';
 import { Post } from '@components';
-import { fetchWithRetry } from '@utils/fetch';
-import logout from '@utils/logout';
-
-async function getPost(postid: number): Promise<PostType[] | undefined> {
-  const url = `/api/posts${postid && `?after=${postid}`}`;
-  const options = {
-    method: 'get',
-  };
-  const response = await fetchWithRetry(url, options);
-  if (response!.status === 200) return await response!.json();
-  else logout();
-}
+import { loadPost } from '@utils/post';
 
 export default function Posts() {
   const [data, setData] = useState<PostType[]>([]);
@@ -19,17 +8,8 @@ export default function Posts() {
   const [loading, setLoading] = useState(false);
   const triggerRef = useRef(null);
 
-  async function loadPost() {
-    setLoading(true);
-    const posts = (await getPost(lastPostId)) as PostType[];
-    console.log(posts);
-    setData([...data, ...posts]);
-    setLoading(false);
-  }
-
   useEffect(() => {
-    loadPost();
-
+    loadPost(setLoading, lastPostId, data, setData);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [lastPostId]);
 
@@ -43,18 +23,17 @@ export default function Posts() {
       },
       { threshold: 1.0 }
     );
+
     const ref = triggerRef.current;
-    if (ref) {
-      observer.observe(ref);
-    }
+    if (ref) observer.observe(ref);
+
     return () => {
-      if (ref) {
-        observer.unobserve(ref);
-      }
+      if (ref) observer.unobserve(ref);
     };
 
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [loading]);
+
   return (
     <>
       {data.map((post, index) => (
